test: 수익률, 당첨개수 맞는지 테스트

+++ src/main/kotlin/lotto/domain/LottoStore.kt
@@ -4,7 +4,7 @@ class LottoStore(private val money: Int, private val lottoMaker: LottoMaker = Lo
     val lottoCount = money / EACH_LOTTO_PRICE
     val boughtLottos = mutableListOf<List<Int>>()
     val totalYieldRatio
-        get() = totalMoney.toFloat() / money
+        get() = totalMoney.toDouble() / money
     private var totalMoney = 0
 
     init {
